refactor: domain 폴더 구조 네이밍 변경
