+++ src/domain/lotto/Lotto.js
