+++ src/domain/lotto/WinningLotto.js
